+++ src/main/java/com/sparta/limited/auth_service/application/service/AuthService.java
@@ -2,9 +2,9 @@ package com.sparta.limited.auth_service.application.service;
 
 import com.sparta.limited.auth_service.application.dto.request.AuthSignupRequest;
 import com.sparta.limited.auth_service.application.dto.response.AuthSignupResponse;
+import com.sparta.limited.auth_service.application.mapper.AuthDtoMapper;
 import com.sparta.limited.auth_service.infrastructure.dto.request.UserCreateInternalRequest;
 import com.sparta.limited.auth_service.infrastructure.dto.response.UserCreateInternalResponse;
-import com.sparta.limited.auth_service.infrastructure.feign.AuthDtoMapper;
 import com.sparta.limited.auth_service.infrastructure.feign.UserFeignClient;
 import com.sparta.limited.auth_service.infrastructure.password.PasswordEncoderUtil;
 import lombok.RequiredArgsConstructor;
